+++ programming_runs/generators/generator_utils.py
@@ -1,12 +1,6 @@
 import os
 from generators.model import ModelBase
-import openai
 import random
-from tenacity import (
-    retry,
-    stop_after_attempt,  # type: ignore
-    wait_random_exponential,  # type: ignore
-)
 
 from typing import Union, List, Optional, Callable
 
@@ -142,60 +136,6 @@ def generic_generate_self_reflection(
         reflection = model.generate(
             f'{SELF_REFLECTION_COMPLETION_INSTRUCTION}\n{func}\n\n{feedback}\n\nExplanation:')
     return reflection  # type: ignore
-
-
-@retry(wait=wait_random_exponential(min=1, max=60), stop=stop_after_attempt(6))
-def gpt_completion(
-        model: str,
-        prompt: str,
-        max_tokens: int = 1024,
-        stop_strs: Optional[List[str]] = None,
-        temperature: float = 0.0,
-        num_comps=1,
-) -> Union[List[str], str]:
-    response = openai.Completion.create(
-        model=model,
-        prompt=prompt,
-        temperature=temperature,
-        max_tokens=max_tokens,
-        top_p=1,
-        frequency_penalty=0.0,
-        presence_penalty=0.0,
-        stop=stop_strs,
-        n=num_comps,
-    )
-    if num_comps == 1:
-        return response.choices[0].text  # type: ignore
-
-    return [choice.text for choice in response.choices]  # type: ignore
-
-
-@retry(wait=wait_random_exponential(min=1, max=180), stop=stop_after_attempt(6))
-def gpt_chat(
-    model: str,
-    system_message: str,
-    user_message: str,
-    max_tokens: int = 1024,
-    temperature: float = 0.0,
-    num_comps=1,
-) -> Union[List[str], str]:
-    response = openai.ChatCompletion.create(
-        model=model,
-        messages=[
-            {"role": "system", "content": system_message},
-            {"role": "user", "content": user_message}
-        ],
-        max_tokens=max_tokens,
-        temperature=temperature,
-        top_p=1,
-        frequency_penalty=0.0,
-        presence_penalty=0.0,
-        n=num_comps,
-    )
-    if num_comps == 1:
-        return response.choices[0].message.content  # type: ignore
-
-    return [choice.message.content for choice in response.choices]  # type: ignore
 
 
 def sample_n_random(items: List[str], n: int) -> List[str]:
